pyproject.toml
[tool.poetry]
name = "auto-dev-supervisor"
version = "0.1.0"
description = "Autonomous Developer Supervisor Application"
authors = ["AutoDev Team"]
readme = "README.md"
packages = [{include = "auto_dev_supervisor", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic = "^2.5.0"
docker = "^7.0.0"
gitpython = "^3.1.0"
pyyaml = "^6.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.12.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
auto-dev = "auto_dev_supervisor.main:app"
